% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_lake.R
\name{daily_lake}
\alias{daily_lake}
\title{daily lake}
\usage{
daily_lake(hypoP = 100, epiP = 100, hypoN = 100, epiN = 100,
  hypoC = 100, epiC = 100, windday, daylength, cyanopop, greenpop, rgreen,
  rcyano, windthresh)
}
\arguments{
\item{hypoP}{(arbitrary units) P concentrations in hypoliminion}

\item{epiP}{(arbitrary units) P concentrations in epiliminion}

\item{hypoN}{(arbitrary units) N concentrations in hypoliminion}

\item{epiN}{(arbitrary units) N concentrations in epiliminion}

\item{hypoC}{(arbitrary units) C concentrations in hypoliminion}

\item{epiC}{(arbitrary units) C concentrations in epiliminion}

\item{windbydayjitter}{(arbitrary units) daily wind}

\item{daylengthbynum}{(hours) daily hours of light}

\item{windthreh}{(arbitrary units) limit of threshold to mix the lake}
}
\value{
daily nutrients and algae populations (arbitrary units)
}
\description{
this function calculates the daily nurient budget and growth of algae populations in the lake the model assumes a one dimensional stratified lake the process takes place in the epilimnion, where growing algae consume the nutrient whan a certain wind threshold is reached, the lake mixes and the nutrients are redistributed between the epi- and the hypolimnion
}
\author{
Yoav BD
}

