% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylengthsimulator.R
\name{daylengthsimulator}
\alias{daylengthsimulator}
\title{day length simulator}
\usage{
daylengthsimulator(mindaylength = 10, maxdaylength = 14, hemisphere = "N")
}
\arguments{
\item{mindaylength}{(hr) minimum day length}

\item{maxdaylength}{(hr) maximum day length}

\item{hemisphere}{(days) define the hemisphere in which the calculation takes place. default value is N, corresponding to 356 (corresponding to DEC 21st)}
}
\value{
daily wind for one year (arbitrary units)
}
\description{
this function simulates a day length in a specified location the user is asked to provide a minimum and maximum day length the values are arranged from JAN 1st to DEC 31st
}
\author{
Yoav BD
}

