% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsimulator.R
\name{windsimulator}
\alias{windsimulator}
\title{wind simulator}
\usage{
windsimulator(cycles = 1, amp = 1, jitfactor = 0.1, jitamount = 0.1,
  phase = 0)
}
\arguments{
\item{cycles}{(-) number of cycles used to model wind (i.e. # of windy seasons per year). default value is 1}

\item{amp}{(-) amplitude of the wind factor (the wind will start from 0 to amp). default value is 1}

\item{jitfactor}{(-) factor of jit, default value is 0.1 of amp}

\item{jitamount}{(-) amount of jit, default value is 0.1 of amp}

\item{phase}{(days) define when the yearly wind cycle starts in days. default value is 0 (corresponding to Jan 1st)}
}
\value{
daily wind for one year (arbitrary units)
}
\description{
this function simulates a daily wind for the daily_lake model the user is asked to provide a factor defining the amplitude of the sine wave, as well as factors defining the jitter (noise) of the wind
}
\author{
Yoav BD
}

